refactor(provisioner): split factory.go into per-distribution files
factory.go held the DefaultFactory creators for all six distributions plus
shared Kubernetes-infra helpers in one 1039-line file. Split each
distribution's creators into its own file (factory_kind.go, factory_k3d.go,
factory_talos.go, factory_vcluster.go, factory_kwok.go, factory_eks.go) with
the cross-distribution host-cluster/DinD helpers in factory_shared.go, leaving
factory.go as the config types, Factory interface and Create dispatch.
Behavior-preserving move via declaration-boundary slicing; per-file imports
rebuilt with goimports.
