+++ pkg/svc/provisioner/cluster/factory.go
@@ -4,34 +4,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
 
 	"github.com/devantler-tech/ksail/v7/pkg/apis/cluster/v1alpha1"
-	eksctlclient "github.com/devantler-tech/ksail/v7/pkg/client/eksctl"
-	"github.com/devantler-tech/ksail/v7/pkg/fsutil"
-	k3dconfigmanager "github.com/devantler-tech/ksail/v7/pkg/fsutil/configmanager/k3d"
-	kindconfigmanager "github.com/devantler-tech/ksail/v7/pkg/fsutil/configmanager/kind"
 	talosconfigmanager "github.com/devantler-tech/ksail/v7/pkg/fsutil/configmanager/talos"
-	"github.com/devantler-tech/ksail/v7/pkg/k8s"
 	"github.com/devantler-tech/ksail/v7/pkg/svc/detector"
-	awsprovider "github.com/devantler-tech/ksail/v7/pkg/svc/provider/aws"
-	kubernetesprovider "github.com/devantler-tech/ksail/v7/pkg/svc/provider/kubernetes"
 	"github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/clustererr"
-	eksprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/eks"
-	k3dprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/k3d"
-	kindprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/kind"
-	kwokprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/kwok"
-	talosprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/talos"
-	vclusterprovisioner "github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/cluster/vcluster"
 	"github.com/devantler-tech/ksail/v7/pkg/svc/provisioner/registry"
 	k3dv1alpha5 "github.com/k3d-io/k3d/v5/pkg/config/v1alpha5"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 	"sigs.k8s.io/kind/pkg/apis/config/v1alpha4"
-	"sigs.k8s.io/yaml"
 )
 
 // Re-export errors for backward compatibility.
@@ -179,861 +159,4 @@ func (f DefaultFactory) Create(
 			cluster.Spec.Cluster.Distribution,
 		)
 	}
-}
-
-func (f DefaultFactory) createKindProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	if f.DistributionConfig.Kind == nil {
-		return nil, nil, fmt.Errorf(
-			"kind config is required for Vanilla distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	kindConfig := f.DistributionConfig.Kind
-
-	// Apply node count overrides from CLI flags / cluster-level config.
-	applyKindNodeCounts(
-		kindConfig,
-		cluster.Spec.Cluster.ControlPlanes,
-		cluster.Spec.Cluster.Workers,
-	)
-
-	// Enable the MutatingAdmissionPolicy feature gate / v1beta1 admissionregistration
-	// API only for Calico, whose v3.30+ CRD chart ships MutatingAdmissionPolicy
-	// resources. Enabling it elsewhere makes other components (e.g. Kyverno) attempt to
-	// use the API and fail.
-	if cluster.Spec.Cluster.CNI == v1alpha1.CNICalico {
-		kindconfigmanager.ApplyAPIServerFeatureGates(kindConfig)
-	}
-
-	// Apply kubelet certificate rotation patches when metrics-server is enabled.
-	// This must happen AFTER applyKindNodeCounts since that function may replace the nodes slice.
-	if cluster.Spec.Cluster.MetricsServer == v1alpha1.MetricsServerEnabled {
-		kindconfigmanager.ApplyKubeletCertRotationPatches(kindConfig)
-	}
-
-	// Apply containerd image verifier plugin patch when image verification is enabled.
-	if cluster.Spec.Cluster.Talos.ImageVerification == v1alpha1.ImageVerificationEnabled {
-		kindconfigmanager.ApplyImageVerificationPatches(kindConfig)
-	}
-
-	// Apply containerd CDI patch when CDI is enabled.
-	cdiVal := cluster.Spec.Cluster.CDI.EffectiveValue(
-		cluster.Spec.Cluster.Distribution, cluster.Spec.Cluster.Provider,
-	)
-	if cdiVal == v1alpha1.CDIEnabled {
-		kindconfigmanager.ApplyCDIPatches(kindConfig)
-	}
-
-	// Kubernetes provider: run Kind inside a DinD pod on the host cluster
-	if cluster.Spec.Cluster.Provider == v1alpha1.ProviderKubernetes {
-		return f.createKindKubernetesProvisioner(cluster, kindConfig)
-	}
-
-	provisioner, err := kindprovisioner.CreateProvisioner(
-		kindConfig,
-		cluster.Spec.Cluster.Connection.Kubeconfig,
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create Kind provisioner: %w", err)
-	}
-
-	if f.ComponentDetector != nil {
-		provisioner.WithComponentDetector(f.ComponentDetector)
-	}
-
-	return provisioner, kindConfig, nil
-}
-
-// createKindKubernetesProvisioner creates a Kind provisioner that runs inside
-// a DinD pod on a host Kubernetes cluster.
-func (f DefaultFactory) createKindKubernetesProvisioner(
-	cluster *v1alpha1.Cluster,
-	kindConfig *v1alpha4.Cluster,
-) (Provisioner, any, error) {
-	opts := cluster.Spec.Provider.Kubernetes
-
-	// Use kindConfig.Name — it's set by applyClusterNameOverride,
-	// while cluster.Name may be empty with --name flag.
-	clusterName := kindConfig.Name
-
-	_, restConfig, dynClient, k8sProvider, err := buildKubernetesInfra(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Configure Kind for DinD: API server must bind to all interfaces
-	// and use a fixed port so it's accessible from outside the DinD container.
-	applyKindDinDNetworking(kindConfig)
-
-	provisioner, err := kindprovisioner.NewKubernetesProvisioner(
-		kindprovisioner.KubernetesProvisionerConfig{
-			KindConfig:       kindConfig,
-			KubeconfigPath:   cluster.Spec.Cluster.Connection.Kubeconfig,
-			K8sProvider:      k8sProvider,
-			DynamicClient:    dynClient,
-			RestConfig:       restConfig,
-			ClusterName:      clusterName,
-			Distribution:     string(cluster.Spec.Cluster.Distribution),
-			GatewayClassName: opts.GatewayClassName,
-			HostContext:      resolveKubernetesOption(opts.Context, opts.ContextEnvVar),
-			APIServerPort:    kubernetesprovider.DinDAPIServerPort,
-			Persistence:      opts.Persistence,
-		},
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create Kind Kubernetes provisioner: %w", err)
-	}
-
-	if f.ComponentDetector != nil {
-		provisioner.WithComponentDetector(f.ComponentDetector)
-	}
-
-	return provisioner, kindConfig, nil
-}
-
-// applyKindDinDNetworking configures Kind networking for DinD execution.
-// The API server must bind to 0.0.0.0 (all interfaces) so it's accessible
-// from outside the DinD container. A fixed port (6443) is used instead of
-// a random port to enable deterministic port mapping. The certSANs patch
-// ensures the API server TLS certificate includes 127.0.0.1 and localhost,
-// which are needed when accessing the API via port-forward.
-func applyKindDinDNetworking(kindConfig *v1alpha4.Cluster) {
-	kindConfig.Networking.APIServerAddress = "0.0.0.0"
-	kindConfig.Networking.APIServerPort = kubernetesprovider.DinDAPIServerPort
-
-	// Add certSANs to all control-plane nodes so the API server cert
-	// is valid for 127.0.0.1 (port-forward) and localhost.
-	certSANsPatch := `kind: ClusterConfiguration
-apiServer:
-  certSANs:
-  - "127.0.0.1"
-  - "localhost"
-  - "0.0.0.0"`
-
-	for i, node := range kindConfig.Nodes {
-		if node.Role == v1alpha4.ControlPlaneRole {
-			kindConfig.Nodes[i].KubeadmConfigPatches = append(
-				kindConfig.Nodes[i].KubeadmConfigPatches,
-				certSANsPatch,
-			)
-		}
-	}
-	// If no nodes are defined, add one with the patch
-	if len(kindConfig.Nodes) == 0 {
-		kindConfig.Nodes = []v1alpha4.Node{
-			{
-				Role:                 v1alpha4.ControlPlaneRole,
-				KubeadmConfigPatches: []string{certSANsPatch},
-			},
-		}
-	}
-}
-
-// wrapK3kServerArgs works around a k3k operator bug: it joins Cluster.Spec.ServerArgs
-// with spaces into an unquoted shell assignment (EXTRA_ARGS={{.EXTRA_ARGS}}), which
-// breaks for multiple space-separated args (the shell treats only the first token as the
-// assignment and the rest as a command). Wrapping the args in a single single-quoted
-// element makes the assignment see one token, while the operator's later unquoted
-// `$EXTRA_ARGS` expansion word-splits it back into separate k3s flags. Returns nil for
-// empty input.
-func wrapK3kServerArgs(args []string) []string {
-	if len(args) == 0 {
-		return nil
-	}
-
-	return []string{"'" + strings.Join(args, " ") + "'"}
-}
-
-// createK3dKubernetesProvisioner creates a K3s provisioner that runs inside
-// a host Kubernetes cluster using the k3k operator.
-func (f DefaultFactory) createK3dKubernetesProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	opts := cluster.Spec.Provider.Kubernetes
-
-	// resolveClusterNameFromContext normally extracts from k3dConfig.Name, but k3d
-	// config is skipped for the Kubernetes provider path. Derive from the connection
-	// context (set by applyClusterNameOverride). The Kubernetes provider writes a
-	// "k3k-<name>" context (via the k3k operator); standalone paths use "k3d-<name>".
-	// Strip whichever prefix is present so the k3k cluster/namespace name is correct.
-	clusterName := strings.TrimPrefix(cluster.Spec.Cluster.Connection.Context, "k3k-")
-	clusterName = strings.TrimPrefix(clusterName, "k3d-")
-
-	if clusterName == "" {
-		clusterName = cluster.Name
-	}
-
-	hostClient, restConfig, dynClient, k8sProvider, err := buildKubernetesInfra(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	controlPlanes := cluster.Spec.Cluster.ControlPlanes
-	if controlPlanes <= 0 {
-		controlPlanes = 1
-	}
-
-	workers := cluster.Spec.Cluster.Workers
-
-	// Calico's v3.30+ CRD chart ships MutatingAdmissionPolicy resources that require the
-	// MutatingAdmissionPolicy feature gate / v1beta1 admissionregistration API. These flags
-	// flow through the k3k Cluster spec's serverArgs to the embedded k3s kube-apiserver.
-	//
-	// The k3k operator joins serverArgs with spaces into an UNQUOTED shell assignment
-	// (EXTRA_ARGS={{.EXTRA_ARGS}}), which breaks for multiple space-separated args (the
-	// shell treats only the first token as the assignment and the rest as a command), so
-	// the feature-gate flags never reach `k3s server $EXTRA_ARGS`. Wrap them in a single
-	// single-quoted element: the assignment then sees one token (EXTRA_ARGS='a b' → "a b")
-	// and the later unquoted $EXTRA_ARGS expansion word-splits it back into separate flags.
-	serverArgs := wrapK3kServerArgs(
-		k3dconfigmanager.APIServerFeatureGatesArgsForCNI(cluster.Spec.Cluster.CNI),
-	)
-
-	provisioner, err := k3dprovisioner.NewK3kProvisioner(
-		k3dprovisioner.K3kProvisionerConfig{
-			HostClientset:    hostClient,
-			RestConfig:       restConfig,
-			K8sProvider:      k8sProvider,
-			DynamicClient:    dynClient,
-			ClusterName:      clusterName,
-			KubeconfigPath:   cluster.Spec.Cluster.Connection.Kubeconfig,
-			HostContext:      resolveKubernetesOption(opts.Context, opts.ContextEnvVar),
-			GatewayClassName: opts.GatewayClassName,
-			ControlPlanes:    controlPlanes,
-			Workers:          workers,
-			PodCIDR:          opts.PodCIDR,
-			ServiceCIDR:      opts.ServiceCIDR,
-			ServerArgs:       serverArgs,
-			// Pin the nested K3s version to the standalone K3d version so the nested
-			// apiserver matches the proven standalone config (serves the
-			// admissionregistration.k8s.io/v1beta1 API that Calico's CRD chart needs),
-			// instead of inheriting the host cluster's possibly-older version.
-			K3sVersion: k3dconfigmanager.DefaultK3sVersion(),
-		},
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create K3k provisioner: %w", err)
-	}
-
-	if f.ComponentDetector != nil {
-		provisioner.WithComponentDetector(f.ComponentDetector)
-	}
-
-	return provisioner, nil, nil
-}
-
-// buildHostClusterClients builds Kubernetes clients for the host cluster
-// from the Kubernetes provider options.
-func buildHostClusterClients(
-	opts v1alpha1.OptionsKubernetes,
-) (kubernetes.Interface, *rest.Config, dynamic.Interface, error) {
-	kubeconfig := resolveKubernetesOption(opts.Kubeconfig, opts.KubeconfigEnvVar)
-
-	// When no kubeconfig is configured, prefer the in-cluster service account if we are
-	// running inside a pod (e.g. the KSail operator). rest.InClusterConfig returns an error
-	// when not in a pod, so CLI behavior outside a cluster is unchanged.
-	if kubeconfig == "" {
-		inClusterConfig, inClusterErr := rest.InClusterConfig()
-		if inClusterErr == nil {
-			return clientsFromRESTConfig(inClusterConfig)
-		}
-
-		kubeconfig = k8s.DefaultKubeconfigPath()
-	}
-
-	// Canonicalize the path (expand ~, resolve symlinks)
-	kubeconfig, err := fsutil.ExpandHomePath(kubeconfig)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("expand kubeconfig path: %w", err)
-	}
-
-	kubeconfig, err = fsutil.EvalCanonicalPath(kubeconfig)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("canonicalize kubeconfig path: %w", err)
-	}
-
-	context := resolveKubernetesOption(opts.Context, opts.ContextEnvVar)
-
-	restConfig, err := k8s.BuildRESTConfig(kubeconfig, context)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("build host REST config: %w", err)
-	}
-
-	return clientsFromRESTConfig(restConfig)
-}
-
-// hostClientQPS and hostClientBurst raise the host-cluster REST client throughput above
-// client-go's restrictive defaults (5 QPS / 10 burst). The nested-provider provisioners
-// poll the host cluster heavily — k3k Cluster status, vCluster kubeconfig secrets, pods —
-// across several nested clusters, and the default rate limiter deadlines under load
-// ("client rate limiter Wait returned an error: context deadline exceeded").
-const (
-	hostClientQPS   = 50
-	hostClientBurst = 100
-)
-
-// clientsFromRESTConfig builds the typed and dynamic clients used by the Kubernetes-provider
-// provisioners from an already-resolved REST config. It raises the client rate-limiter
-// throughput (unless already configured) so readiness polling survives host-cluster load.
-func clientsFromRESTConfig(
-	restConfig *rest.Config,
-) (kubernetes.Interface, *rest.Config, dynamic.Interface, error) {
-	if restConfig.QPS == 0 {
-		restConfig.QPS = hostClientQPS
-	}
-
-	if restConfig.Burst == 0 {
-		restConfig.Burst = hostClientBurst
-	}
-
-	clientset, err := kubernetes.NewForConfig(restConfig)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("create host clientset: %w", err)
-	}
-
-	dynClient, err := dynamic.NewForConfig(restConfig)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("create host dynamic client: %w", err)
-	}
-
-	return clientset, restConfig, dynClient, nil
-}
-
-// resolveKubernetesOption returns the value from the environment variable
-// if set, otherwise returns the direct value.
-func resolveKubernetesOption(directValue, envVar string) string {
-	if envVar != "" {
-		if envValue := os.Getenv(envVar); envValue != "" {
-			return envValue
-		}
-	}
-
-	return directValue
-}
-
-// buildKubernetesInfra creates the host-cluster clients and Kubernetes provider in one call.
-// This consolidates the repeated buildHostClusterClients + NewProvider sequence.
-func buildKubernetesInfra(
-	opts v1alpha1.OptionsKubernetes,
-) (kubernetes.Interface, *rest.Config, dynamic.Interface, *kubernetesprovider.Provider, error) {
-	hostClient, restConfig, dynClient, err := buildHostClusterClients(opts)
-	if err != nil {
-		return nil, nil, nil, nil, fmt.Errorf("build host cluster clients: %w", err)
-	}
-
-	k8sProvider, err := kubernetesprovider.NewProvider(hostClient, opts)
-	if err != nil {
-		return nil, nil, nil, nil, fmt.Errorf("create Kubernetes provider: %w", err)
-	}
-
-	return hostClient, restConfig, dynClient, k8sProvider, nil
-}
-
-// applyKindNodeCounts applies node count overrides from CLI flags / cluster-level
-// config to the Kind config. Enables --control-planes and --workers CLI flags to
-// override kind.yaml at runtime.
-func applyKindNodeCounts(kindConfig *v1alpha4.Cluster, controlPlanes, workers int32) {
-	if controlPlanes <= 0 && workers <= 0 {
-		return
-	}
-
-	targetCP := int(controlPlanes)
-	if targetCP <= 0 {
-		targetCP = 1 // default to 1 control-plane
-	}
-
-	targetWorkers := int(workers)
-
-	newNodes := make([]v1alpha4.Node, 0, targetCP+targetWorkers)
-
-	for range targetCP {
-		newNodes = append(newNodes, v1alpha4.Node{
-			Role:  v1alpha4.ControlPlaneRole,
-			Image: kindconfigmanager.DefaultKindNodeImage,
-		})
-	}
-
-	for range targetWorkers {
-		newNodes = append(newNodes, v1alpha4.Node{
-			Role:  v1alpha4.WorkerRole,
-			Image: kindconfigmanager.DefaultKindNodeImage,
-		})
-	}
-
-	kindConfig.Nodes = newNodes
-}
-
-func (f DefaultFactory) createK3dProvisioner( //nolint:funlen // sequential setup steps
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	// Kubernetes provider: use k3k operator instead of Docker-based K3d
-	if cluster.Spec.Cluster.Provider == v1alpha1.ProviderKubernetes {
-		return f.createK3dKubernetesProvisioner(cluster)
-	}
-
-	if f.DistributionConfig.K3d == nil {
-		return nil, nil, fmt.Errorf(
-			"k3d config is required for K3d distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	k3dConfig := f.DistributionConfig.K3d
-
-	// Apply node count overrides from CLI flags / cluster-level config.
-	applyK3dNodeCounts(k3dConfig, cluster.Spec.Cluster.ControlPlanes, cluster.Spec.Cluster.Workers)
-
-	// Enable the MutatingAdmissionPolicy feature gate / v1beta1 admissionregistration
-	// API on the server nodes only for Calico, whose v3.30+ CRD chart ships
-	// MutatingAdmissionPolicy resources. Enabling it elsewhere makes other components
-	// (e.g. Kyverno) attempt to use the API and fail.
-	if cluster.Spec.Cluster.CNI == v1alpha1.CNICalico {
-		k3dconfigmanager.ApplyAPIServerFeatureGatesArgs(k3dConfig)
-	}
-
-	// Apply containerd image verifier plugin volume mount when image verification is enabled.
-	// This mounts the generated config.toml.tmpl into K3d node containers so K3s uses it
-	// to generate the final containerd config with the image verifier plugin enabled.
-	if cluster.Spec.Cluster.Talos.ImageVerification == v1alpha1.ImageVerificationEnabled {
-		templatePath := filepath.Join(k3dconfigmanager.DefaultImageVerifierDir, "config.toml.tmpl")
-
-		absTemplatePath, err := filepath.Abs(templatePath)
-		if err != nil {
-			return nil, nil, fmt.Errorf(
-				"failed to resolve k3d image verification template path %q: %w",
-				templatePath,
-				err,
-			)
-		}
-
-		fileInfo, err := os.Stat(absTemplatePath)
-		if err != nil {
-			return nil, nil, fmt.Errorf(
-				"k3d image verification template not found at %q; run 'ksail cluster init' to generate it: %w",
-				absTemplatePath,
-				err,
-			)
-		}
-
-		if !fileInfo.Mode().IsRegular() {
-			return nil, nil, fmt.Errorf(
-				"%w: %s; remove it and re-run 'ksail cluster init'",
-				ErrImageVerificationTemplateNotRegularFile,
-				absTemplatePath,
-			)
-		}
-
-		k3dconfigmanager.ApplyImageVerificationVolumes(k3dConfig, absTemplatePath)
-	}
-
-	// Write the in-memory config to a temp file so k3d picks up any modifications
-	// (e.g., registry mirrors configured via --mirror-registry, node counts).
-	// We always use a temp file to avoid modifying the user's k3d.yaml.
-	// The k3d CLI reads configuration from file, not from our in-memory config.
-	tempConfigPath, err := writeK3dConfigToTempFile(k3dConfig)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to write k3d config to temp file: %w", err)
-	}
-
-	provisioner := k3dprovisioner.CreateProvisioner(
-		k3dConfig,
-		tempConfigPath,
-	).WithKubeconfig(cluster.Spec.Cluster.Connection.Kubeconfig)
-
-	if f.ComponentDetector != nil {
-		provisioner.WithComponentDetector(f.ComponentDetector)
-	}
-
-	return provisioner, k3dConfig, nil
-}
-
-// applyK3dNodeCounts applies node count overrides from CLI flags / cluster-level
-// config to the K3d config. Enables --control-planes and --workers CLI flags to
-// override k3d.yaml at runtime.
-func applyK3dNodeCounts(k3dConfig *k3dv1alpha5.SimpleConfig, controlPlanes, workers int32) {
-	if controlPlanes <= 0 && workers <= 0 {
-		return
-	}
-
-	if controlPlanes > 0 {
-		k3dConfig.Servers = int(controlPlanes)
-	}
-
-	k3dConfig.Agents = int(workers)
-}
-
-// talosOptsFromCluster builds the OptionsTalos passed to the Talos provisioner,
-// overlaying the distribution-agnostic cluster fields (node counts, Kubernetes
-// version) that the provisioner consumes. The node-count fields bridge the
-// deprecated Talos-scoped aliases during the migration window; KubernetesVersion
-// carries the top-level spec.cluster.kubernetesVersion pin so the provisioner can
-// tell whether the user pinned a version (vs. tracking the running one on update).
-func talosOptsFromCluster(cluster *v1alpha1.Cluster) v1alpha1.OptionsTalos {
-	talosOpts := cluster.Spec.Cluster.Talos
-	//nolint:staticcheck // intentional: bridging deprecated field
-	talosOpts.ControlPlanes = cluster.Spec.Cluster.ControlPlanes
-	//nolint:staticcheck // intentional: bridging deprecated field
-	talosOpts.Workers = cluster.Spec.Cluster.Workers
-	talosOpts.KubernetesVersion = cluster.Spec.Cluster.KubernetesVersion
-
-	return talosOpts
-}
-
-func (f DefaultFactory) createTalosProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	// Kubernetes provider: run Talos inside a DinD pod on the host cluster
-	if cluster.Spec.Cluster.Provider == v1alpha1.ProviderKubernetes {
-		return f.createTalosKubernetesProvisioner(cluster)
-	}
-
-	if f.DistributionConfig.Talos == nil {
-		return nil, nil, fmt.Errorf(
-			"talos config is required for Talos distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	// Always skip CNI-dependent checks (CoreDNS, kube-proxy) for Talos Docker provisioner.
-	//
-	// Rationale:
-	// 1. Custom CNI (Cilium, Calico): Pods cannot start until CNI is installed post-bootstrap.
-	// 2. Default Flannel CNI: While Flannel is bundled with Talos, it can be slow or unreliable
-	//    in containerized environments (GitHub Actions, Docker-in-Docker). The checks for
-	//    kube-proxy and CoreDNS can timeout even when the cluster is fundamentally healthy.
-	//
-	// Since we've verified that etcd, kubelet, and the Kubernetes API are healthy via
-	// PreBootSequenceChecks, the cluster is functional. Application-level DNS/proxy
-	// services will become ready shortly after bootstrap completes.
-	skipCNIChecks := true
-
-	// Overlay cluster-level fields (node counts, Kubernetes version) onto Talos options.
-	talosOpts := talosOptsFromCluster(cluster)
-
-	// Propagate autoscaler-enabled flag to Hetzner options so the provisioner
-	// can create the cluster-autoscaler-config Secret during bootstrap.
-	hetznerOpts := cluster.Spec.Provider.Hetzner
-
-	hetznerOpts.NodeAutoscalerEnabled = cluster.Spec.Cluster.Autoscaler.Node.Enabled ||
-		cluster.Spec.Cluster.NodeAutoscaling == v1alpha1.NodeAutoscalingEnabled
-
-	// Derive pool names from the new autoscaler pools config so that the
-	// delete path can clean up autoscaler-managed Hetzner servers.
-	if len(hetznerOpts.AutoscalerNodePoolNames) == 0 {
-		pools := cluster.Spec.Cluster.Autoscaler.Node.Pools
-		if len(pools) > 0 {
-			names := make([]string, len(pools))
-			for i, pool := range pools {
-				names[i] = pool.Name
-			}
-
-			hetznerOpts.AutoscalerNodePoolNames = names
-		}
-	}
-
-	provisioner, err := talosprovisioner.CreateProvisioner(
-		f.DistributionConfig.Talos,
-		cluster.Spec.Cluster.Connection.Kubeconfig,
-		cluster.Spec.Cluster.Connection.Context,
-		cluster.Spec.Cluster.Provider,
-		talosOpts,
-		hetznerOpts,
-		cluster.Spec.Provider.Omni,
-		skipCNIChecks,
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create Talos provisioner: %w", err)
-	}
-
-	if f.ComponentDetector != nil {
-		provisioner.WithComponentDetector(f.ComponentDetector)
-	}
-
-	return provisioner, f.DistributionConfig.Talos, nil
-}
-
-// createTalosKubernetesProvisioner creates a Talos provisioner that runs inside
-// a DinD pod on a host Kubernetes cluster via the Talos SDK.
-func (f DefaultFactory) createTalosKubernetesProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	if f.DistributionConfig.Talos == nil {
-		return nil, nil, fmt.Errorf(
-			"talos config is required for Talos distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	opts := cluster.Spec.Provider.Kubernetes
-
-	// Derive cluster name from Talos config (set by applyClusterNameOverride).
-	clusterName := f.DistributionConfig.Talos.GetClusterName()
-	if clusterName == "" {
-		clusterName = cluster.Name
-	}
-
-	_, restConfig, dynClient, k8sProvider, err := buildKubernetesInfra(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Create a full inner Talos Provisioner (Docker provider type).
-	// The Docker client will be injected at Create() time after DinD is ready.
-	talosOpts := talosOptsFromCluster(cluster)
-
-	innerProvisioner, err := talosprovisioner.CreateProvisioner(
-		f.DistributionConfig.Talos,
-		cluster.Spec.Cluster.Connection.Kubeconfig,
-		"",
-		v1alpha1.ProviderDocker,
-		talosOpts,
-		v1alpha1.OptionsHetzner{},
-		v1alpha1.OptionsOmni{},
-		true, // skipCNIChecks — same as normal Docker path
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create inner Talos provisioner: %w", err)
-	}
-
-	// jscpd:ignore-start
-	provisioner, err := talosprovisioner.NewKubernetesProvisioner(
-		talosprovisioner.KubernetesProvisionerConfig{
-			InnerProvisioner: innerProvisioner,
-			KubeconfigPath:   cluster.Spec.Cluster.Connection.Kubeconfig,
-			K8sProvider:      k8sProvider,
-			DynamicClient:    dynClient,
-			RestConfig:       restConfig,
-			ClusterName:      clusterName,
-			Distribution:     string(cluster.Spec.Cluster.Distribution),
-			GatewayClassName: opts.GatewayClassName,
-			HostContext:      resolveKubernetesOption(opts.Context, opts.ContextEnvVar),
-			ControlPlanes:    int(cluster.Spec.Cluster.ControlPlanes),
-			Workers:          int(cluster.Spec.Cluster.Workers),
-			Persistence:      opts.Persistence,
-			MirrorSpecs:      f.DistributionConfig.MirrorSpecs,
-		},
-	)
-	// jscpd:ignore-end
-	if err != nil {
-		return nil, nil, fmt.Errorf("create Talos Kubernetes provisioner: %w", err)
-	}
-
-	return provisioner, nil, nil
-}
-
-func (f DefaultFactory) createVClusterProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	if cluster.Spec.Cluster.Provider == v1alpha1.ProviderKubernetes {
-		return f.createVClusterKubernetesProvisioner(cluster)
-	}
-
-	if f.DistributionConfig.VCluster == nil {
-		return nil, nil, fmt.Errorf(
-			"vcluster config is required for VCluster distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	vclusterConfig := f.DistributionConfig.VCluster
-
-	provisioner, err := vclusterprovisioner.CreateProvisioner(
-		vclusterConfig.Name,
-		vclusterConfig.ValuesPath,
-		vclusterConfig.DisableFlannel,
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create VCluster provisioner: %w", err)
-	}
-
-	return provisioner, vclusterConfig, nil
-}
-
-// createVClusterKubernetesProvisioner creates a vCluster provisioner that
-// deploys vCluster as a Helm release on a host Kubernetes cluster.
-func (f DefaultFactory) createVClusterKubernetesProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	opts := cluster.Spec.Provider.Kubernetes
-
-	// Use VCluster config name (set by applyClusterNameOverride).
-	vclusterConfig := f.DistributionConfig.VCluster
-
-	clusterName := ""
-	if vclusterConfig != nil {
-		clusterName = vclusterConfig.Name
-	}
-
-	if clusterName == "" {
-		clusterName = cluster.Name
-	}
-
-	hostClient, restConfig, dynClient, k8sProvider, err := buildKubernetesInfra(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var (
-		valuesPath     string
-		disableFlannel bool
-	)
-
-	if vclusterConfig != nil {
-		valuesPath = vclusterConfig.ValuesPath
-		disableFlannel = vclusterConfig.DisableFlannel
-	}
-
-	provisioner, err := vclusterprovisioner.NewKubernetesProvisioner(
-		vclusterprovisioner.KubernetesProvisionerConfig{
-			ClusterName:      clusterName,
-			HostContext:      resolveKubernetesOption(opts.Context, opts.ContextEnvVar),
-			KubeconfigPath:   cluster.Spec.Cluster.Connection.Kubeconfig,
-			HostClientset:    hostClient,
-			RestConfig:       restConfig,
-			K8sProvider:      k8sProvider,
-			DynamicClient:    dynClient,
-			GatewayClassName: opts.GatewayClassName,
-			ValuesPath:       valuesPath,
-			DisableFlannel:   disableFlannel,
-		},
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create vCluster Kubernetes provisioner: %w", err)
-	}
-
-	return provisioner, vclusterConfig, nil
-}
-
-func (f DefaultFactory) createKWOKProvisioner(
-	cluster *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	if f.DistributionConfig.KWOK == nil {
-		return nil, nil, fmt.Errorf(
-			"kwok config is required for KWOK distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	kwokConfig := f.DistributionConfig.KWOK
-
-	// Kubernetes provider: run KWOK inside a DinD pod on the host cluster
-	if cluster.Spec.Cluster.Provider == v1alpha1.ProviderKubernetes {
-		return f.createKWOKKubernetesProvisioner(cluster, kwokConfig)
-	}
-
-	provisioner := kwokprovisioner.NewProvisioner(
-		kwokConfig.Name,
-		kwokConfig.ConfigPath,
-		nil,
-	)
-
-	return provisioner, kwokConfig, nil
-}
-
-// createKWOKKubernetesProvisioner creates a KWOK provisioner that runs inside
-// a DinD pod on a host Kubernetes cluster.
-func (f DefaultFactory) createKWOKKubernetesProvisioner(
-	cluster *v1alpha1.Cluster,
-	kwokConfig *KWOKConfig,
-) (Provisioner, any, error) {
-	opts := cluster.Spec.Provider.Kubernetes
-
-	_, restConfig, dynClient, k8sProvider, err := buildKubernetesInfra(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Use kwokConfig.Name as the cluster name — it's always set correctly
-	// by applyClusterNameOverride, while cluster.Name may be empty
-	// when using --name flag without a ksail.yaml file.
-	clusterName := kwokConfig.Name
-
-	provisioner, err := kwokprovisioner.NewKubernetesProvisioner(
-		kwokprovisioner.KubernetesProvisionerConfig{
-			Name:             kwokConfig.Name,
-			ConfigPath:       kwokConfig.ConfigPath,
-			KubeconfigPath:   cluster.Spec.Cluster.Connection.Kubeconfig,
-			K8sProvider:      k8sProvider,
-			DynamicClient:    dynClient,
-			RestConfig:       restConfig,
-			ClusterName:      clusterName,
-			Distribution:     string(cluster.Spec.Cluster.Distribution),
-			GatewayClassName: opts.GatewayClassName,
-			Persistence:      opts.Persistence,
-		},
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("create KWOK Kubernetes provisioner: %w", err)
-	}
-
-	return provisioner, kwokConfig, nil
-}
-
-func (f DefaultFactory) createEKSProvisioner(
-	_ *v1alpha1.Cluster,
-) (Provisioner, any, error) {
-	if f.DistributionConfig.EKS == nil {
-		return nil, nil, fmt.Errorf(
-			"eks config is required for EKS distribution: %w",
-			ErrMissingDistributionConfig,
-		)
-	}
-
-	eksConfig := f.DistributionConfig.EKS
-	client := eksctlclient.NewClient()
-
-	infraProvider, err := awsprovider.NewProvider(client, eksConfig.Region)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create AWS provider: %w", err)
-	}
-
-	provisioner, err := eksprovisioner.NewProvisioner(
-		eksConfig.Name,
-		eksConfig.Region,
-		eksConfig.ConfigPath,
-		client,
-		infraProvider,
-	)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create EKS provisioner: %w", err)
-	}
-
-	return provisioner, eksConfig, nil
-}
-
-// writeK3dConfigToTempFile writes the in-memory k3d config to a temporary file.
-// This approach avoids modifying the user's k3d.yaml while ensuring k3d picks up
-// all in-memory modifications (registry mirrors, node counts, etc.).
-// The temp file persists until system cleanup - this is intentional since k3d
-// may reference the config path during cluster operations.
-func writeK3dConfigToTempFile(config *k3dv1alpha5.SimpleConfig) (string, error) {
-	data, err := yaml.Marshal(config)
-	if err != nil {
-		return "", fmt.Errorf("marshal k3d config: %w", err)
-	}
-
-	// Create temp file with k3d prefix for easy identification
-	tempFile, err := os.CreateTemp("", "ksail-k3d-*.yaml")
-	if err != nil {
-		return "", fmt.Errorf("create temp file: %w", err)
-	}
-
-	filePath := tempFile.Name()
-
-	_, writeErr := tempFile.Write(data)
-
-	closeErr := tempFile.Close()
-
-	if writeErr != nil {
-		return "", fmt.Errorf("write to temp file: %w", writeErr)
-	}
-
-	if closeErr != nil {
-		return "", fmt.Errorf("close temp file: %w", closeErr)
-	}
-
-	return filePath, nil
 }
